+++ internal/cp/proxy_manager_test.go
